+++ Apps/Contoso.Android.Puppet/MainActivity.cs
@@ -66,7 +66,8 @@ namespace Contoso.Android.Puppet
             }
             Distribute.SetInstallUrl("https://install.portal-server-core-integration.dev.avalanch.es");
             Distribute.SetApiUrl("https://asgard-int.trafficmanager.net/api/v0.1");
-            AppCenter.Start("bff0949b-7970-439d-9745-92cdc59b10fe", typeof(Analytics), typeof(Crashes), typeof(Distribute));
+            var appCenterSecret = Intent.GetStringExtra("XAMARIN_ANDROID_INT");
+            AppCenter.Start(appCenterSecret, typeof(Analytics), typeof(Crashes), typeof(Distribute));
             AppCenter.IsEnabledAsync().ContinueWith(enabled =>
             {
                 AppCenterLog.Info(LogTag, "AppCenter.Enabled=" + enabled.Result);