+++ Apps/Contoso.WPF.Puppet.DotNetCore/App.xaml.cs
@@ -48,7 +48,8 @@ namespace Contoso.WPF.Puppet.DotNetCore
             }
 
             // Start AppCenter.
-            AppCenter.Start("d967daf9-28ed-4899-84e8-17a00c064987", typeof(Analytics), typeof(Crashes));
+            var appSecret = Environment.GetEnvironmentVariable("WPF_CORE_INT");
+            AppCenter.Start(appSecret, typeof(Analytics), typeof(Crashes));
             var userId = Settings.Default.UserId;
             if (!string.IsNullOrEmpty(userId))
             {