+++ Apps/Contoso.WinForms.Demo/Program.cs
@@ -34,7 +34,8 @@ namespace Contoso.WinForms.Demo
             {
                 AppCenter.SetMaxStorageSizeAsync(storageMaxSize);
             }
-            AppCenter.Start("734be4f7-3607-489b-ae81-284d2eb908f8", typeof(Analytics), typeof(Crashes));
+            var appSecret = Environment.GetEnvironmentVariable("WINFORMS_PROD");
+            AppCenter.Start(appSecret, typeof(Analytics), typeof(Crashes));
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
             Application.Run(new MainForm());