+++ Apps/Contoso.WinForms.Puppet/Program.cs
@@ -34,7 +34,8 @@ namespace Contoso.WinForms.Puppet
             {
                 AppCenter.SetMaxStorageSizeAsync(storageMaxSize);
             }
-            AppCenter.Start("7136db69-7f8d-4a14-90bd-12c9588ae0b9", typeof(Analytics), typeof(Crashes));
+            var appSecret = Environment.GetEnvironmentVariable("WINFORMS_INT");
+            AppCenter.Start(appSecret, typeof(Analytics), typeof(Crashes));
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
             Application.Run(new MainForm());