+++ Apps/Contoso.WinUI.Desktop.Puppet/Contoso.WinUI.Desktop.Puppet/App.xaml.cs
@@ -73,7 +73,8 @@ namespace Contoso.WinUI.Desktop.Puppet
             Crashes.FailedToSendErrorReport += (_, args) => Log($"Failed to send error report for an error ID: {args.Report.Id}");
 
             // Start App Center.
-            AppCenter.Start("2daf2955-b4e4-4206-b38a-fedcdc6f4f84", typeof(Analytics), typeof(Crashes));
+            var appSecret = Environment.GetEnvironmentVariable("WINUI_IN_DESKTOP_INT");
+            AppCenter.Start(appSecret, typeof(Analytics), typeof(Crashes));
 
             // Set userId.
             var userId = localSettings.Values[Constants.KeyUserId] as string;